Add (commented-out) section on noise model.
Discuss impact of non-white detector noise on image NETD. But getting
this to the point where I really trust my results is too much work for
now, so I'm not ready to really include it in the thesis. Saving it in
comments in case I get more confidence in in later.